backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
diff-match-patch==20241021
Django==5.1.6
django-appconf==1.1.0
django-browser-reload==1.18.0
django-cors-headers==4.0.0
django-debug-toolbar==5.0.1
django-htmx==1.22.0
django-import-export==4.3.6
django-neomodel==0.1.1
django-select2==8.4.0
django-tailwind==3.8.0
django-vite==3.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
neo4j==5.12.0
neobolt==1.7.17
neomodel==5.1.2
packaging==24.2
psycopg==3.2.4
pymarc==5.2.3
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
six==1.16.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
uritemplate==4.1.1
